student-project matching: least popular project removal
Iteratively remove the project that is "least popular". That is, whose removal leads to maximal metric (AUPCR) value compared to any other.

A very brute-force'y approach and slow. GraphMatch has been optimized a little bit and the least-popular search is threaded, but still very slow and result is yet undetermined