Add and use MinQuorum over Min GrpSizeConstraint
Generalizes to sitatuations when the number of agents to add is less than a min viable group size. For example, in valid group sizes of 4-5, when a grp exists of 4 members, we can simply add one more member - that is, if we want to we can now do so.

Used by SDPC